Allow custom splitSentences characters
splitSentences can now be true, false, or set to a string of custom
characters to deliminate the sentences.

An error is thrown if [a-zA-Z] are passed (a user sending in 'true'
could be in for a world of hurt  figuring that out.)

Any number of consecutive splitting characters can be used, and
the sentence will only be split once.